Change mqtt snippet to default data type for length

+++ snippets/mqtt_client.c
@@ -176,9 +176,9 @@ mqtt_cb(mqtt_client_t* client, mqtt_evt_t* evt) {
          */
         case MQTT_EVT_PUBLISH_RECV: {
             const char* topic = (const char *)evt->evt.publish_recv.topic;
-            uint16_t topic_len = (uint16_t)evt->evt.publish_recv.topic_len;
+            size_t topic_len = evt->evt.publish_recv.topic_len;
             const uint8_t* payload = evt->evt.publish_recv.payload;
-            uint16_t payload_len = (uint16_t)evt->evt.publish_recv.payload_len;
+            size_t payload_len = evt->evt.publish_recv.payload_len;
             
             ESP_UNUSED(payload);
             ESP_UNUSED(payload_len);
